+++ src/components/Profile/index.tsx
@@ -1,0 +1,1 @@
+export const hi = "hi";
