+++ src/pages/HomeModal/HomeModal.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import Modal from "../../contents/Modal/Modal";
+import Modal from "../../components/Modal/Modal";
 import {
   ChoiceBox,
   Phrase,
