+++ src/pages/PinPage/index.tsx
@@ -3,7 +3,7 @@ import { styled } from "styled-components";
 import { themeState } from "../../atom/atom";
 import { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import CommonInput from "../../contents/Input/Input";
+import CommonInput from "../../components/Input/Input";
 import { FlexColumnCenterCSS } from "../../styled-components/commonStyle";
 
 function Pinpage() {
